pyproject.toml
[project]
name = "writing-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "google-genai>=1.15.0",
    "google-generativeai>=0.8.5",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
]
